Introduce table clipboard integration plugin.
